Fix #4
Fix a bug that pooling layer is merged with convolution layer that does
not use that layer as input layer.

Fix ZeroPadding layers parsing in Keras so the layer that use it can
correctly get the padding info if the ZeroPadding is not directly
followed by the layer that use it.